♻️ Move language logic to own hook
