+++ frontend/src/components/form-degree.tsx
@@ -1,8 +1,7 @@
-import { useContext } from 'react';
 import type { SelectProps } from '@chakra-ui/react';
 import { Heading, FormControl, FormLabel, Select } from '@chakra-ui/react';
 import { useFormContext } from 'react-hook-form';
-import LanguageContext from 'language-context';
+import useLanguage from '@hooks/use-language';
 
 interface Props extends SelectProps {
     isHeading?: boolean;
@@ -11,7 +10,7 @@ interface Props extends SelectProps {
 }
 const FormDegree = ({ isHeading = false, hideLabel = false, isRequired = false, ...props }: Props) => {
     const { register } = useFormContext();
-    const isNorwegian = useContext(LanguageContext);
+    const isNorwegian = useLanguage();
     const headingText = isNorwegian ? 'Studieretning' : 'Field of study';
 
     return (
