+++ frontend/src/components/minute-list.tsx
@@ -1,4 +1,3 @@
-import { useContext } from 'react';
 import {
     Flex,
     Heading,
@@ -19,7 +18,7 @@ import { nb, enUS } from 'date-fns/locale';
 import NextLink from 'next/link';
 import { FaExternalLinkAlt } from 'react-icons/fa';
 import type { Minute } from '@api/minute';
-import LanguageContext from 'language-context';
+import useLanguage from '@hooks/use-language';
 
 interface Props {
     minutes: Array<Minute>;
@@ -29,7 +28,7 @@ const MinuteList = ({ minutes }: Props): JSX.Element => {
     const newMinutes = minutes.map((obj) => ({ ...obj, date: new Date(obj.date) }));
 
     const color = useColorModeValue('blue', 'blue.400');
-    const isNorwegian = useContext(LanguageContext);
+    const isNorwegian = useLanguage();
     const years = [...new Set(newMinutes.map((e) => e.date.getFullYear()))];
 
     return (
