--- frontend/src/language-context.tsx
@@ -1,4 +1,0 @@
-import { createContext } from 'react';
-const LanguageContext = createContext(true);
-
-export default LanguageContext;
