+++ frontend/src/pages/_app.tsx
@@ -4,35 +4,18 @@ import { getMonth } from 'date-fns';
 import type { AppProps } from 'next/app';
 import { useRouter } from 'next/router';
 import NextNProgress from 'nextjs-progressbar';
-import { useEffect, useState } from 'react';
 import Snowfall from 'react-snowfall';
 import Fonts from '@styles/fonts';
 import theme from '@styles/theme';
 import Layout from '@components/layout';
-import LanguageContext from 'language-context';
+import { LanguageProvider } from '@hooks/use-language';
 
 const App = ({ Component, pageProps: { session, ...pageProps } }: AppProps<SessionProviderProps>): JSX.Element => {
     const router = useRouter();
     const SSR = typeof window === 'undefined'; //Used to disable rendering of animated component SS
-    const [isNorwegian, setIsNorwegian] = useState(true);
-    useEffect(() => {
-        const checkLanguageData = () => {
-            const lang = localStorage.getItem('language');
-            if (lang === 'en') {
-                setIsNorwegian(false);
-            } else {
-                setIsNorwegian(true);
-            }
-        };
-        checkLanguageData();
-        window.addEventListener('storage', checkLanguageData);
-        return () => {
-            window.removeEventListener('storage', checkLanguageData);
-        };
-    }, []);
 
     return (
-        <LanguageContext.Provider value={isNorwegian}>
+        <LanguageProvider>
             <SessionProvider session={session}>
                 <ChakraProvider theme={theme}>
                     <NextNProgress
@@ -49,7 +32,7 @@ const App = ({ Component, pageProps: { session, ...pageProps } }: AppProps<Sessi
                     </Layout>
                 </ChakraProvider>
             </SessionProvider>
-        </LanguageContext.Provider>
+        </LanguageProvider>
     );
 };
